streamlit_app/requirements.txt
streamlit
pandas
numpy
python-pptx
python-dateutil
sqlalchemy
psycopg2-binary
plotly-express